[OBEY-CAMPAIGN-8deed8b4-FE-CW0003] fix(workitem): address commit rereview
